+++ app/services/fulfillment/validate_order.rb
@@ -1,0 +1,4 @@
+module Fulfillment
+  ValidateOrder = Printful::ValidateOrder
+  CreateOrder = Printful::CreateOrder
+end
