--- app/services/printful/create_order.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Printful
-  class CreateOrder < Base
-    def call(confirm: false)
-      payload = purchase_order_payload(include_id: true)
-      PrintfulAPI::Order.create(payload, confirm: confirm)
-    end
-  end
-end
